ENH: Make the remaining integer conversions explicit
Last of five commits clearing -Wsign-conversion: 29 hunks that did not
fall into the allocation, copy or I/O groups.

  File offsets and positions.  znztell() returns a signed znz_off_t and
  its result is stored in size_t locals; iname_offset and the extension
  sizes are ints that take part in size_t arithmetic.

  Extension bookkeeping.  `remain -= len` in the four fill_cmd_string()
  and extension loops, where remain is size_t and len an int already
  known non-negative.

  Chained assignments.  `nbl->bsize = nbl->nbricks = 0` assigns an int 0
  through a size_t field; split into two statements, which is also
  clearer about the two different types.

  The ANALYZE orientation byte.  nifti_convert_nhdr2nim() reads it with
  `unsigned char c = *((char *)(&nhdr.qform_code));`, taking a signed
  char and widening it into an unsigned one -- so any value above 127
  arrives sign-extended.  Reading through unsigned char * instead gives
  the byte that is actually in the header.

  nifti_get_filesize() returned `(unsigned int)buf.st_size` from a
  function declared to return int, converting off_t to unsigned and then
  to signed.  Now a single cast to the declared return type.

With this the tree is free of -Wsign-conversion warnings under both gcc
16.1.1 and clang 22.1.8, in the default, USE_FSL_CODE and FSLSTYLE/PIGZ
configurations, and 92/92 tests pass in all six.

+++ cifti/afni_xml_io.c
@@ -144,7 +144,7 @@ axio_num_tokens(const char * str, int64_t maxlen)
   if (maxlen > 0)
     len = maxlen;
   else
-    len = strlen(str);
+    len = (int64_t)strlen(str);
 
   ntok = 0;
   intok = 0;
